Organize REST API server Go files
